Update Extras-Addon.lua
Tabs are now automatically collapsed instead of opened and can also be resized and moved around to different locations on the screen.

Vehicle Gifting now deletes invisible vehicles for you as long as you are within the range of them AND automatically sends a chat message to the receiving player.  -- Thanks to hexarobi/Holy

Custom Wheel listboxes now show beside one another as to not take up so much space.